# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dtos.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
      remember_me:
        description: RememberMe extends the iam_session cookie lifetime (SSO_SESSION_REMEMBER_TTL).
          API login and POST /oidc/login both honor this.
        type: boolean
      return_to:
        description: 'ReturnTo when set (OIDC browser flow): creates a session cookie
          and redirects here with 302 — no access_token in the response body.'
        type: string
      tenant_id:
        description: TenantID optional explicit tenant context.
        type: string
    required:
    - email
    - password
    type: object
  dtos.LoginResponse:
    properties:
      access_token:
        type: string
      active_tenant_id:
        type: string
      expires_in:
        description: access token lifetime (seconds)
        type: integer
      refresh_expires_in:
        description: refresh token lifetime (seconds)
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
...
  dtos.Meta:
    description: Metadata for pagination
    properties:
      code:
        example: 400
        type: integer
      error_code:
        example: BAD_REQUEST
        type: string
      message:
        example: Request is invalid
        type: string
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      total:
        example: 18
        type: integer
    type: object
...
  dtos.TokenIntrospectionResponse:
    properties:
      active:
        type: boolean
      aud:
        type: string
      client_id:
        type: string
      exp:
        type: integer
      iat:
        type: integer
      iss:
        type: string
      jti:
        type: string
      nbf:
        type: integer
      scope:
        type: string
      sub:
        type: string
      token_type:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: GateForge IAM application API under /api/v1. The canonical public SDK
    contract (including root OIDC paths) lives at api/openapi.yaml in the monorepo.
  title: GateForge IAM API
  version: 0.1.0
paths:
...
  /authorize:
    get:
      consumes:
      - application/json
      parameters:
      - description: Session cookie from browser login (iam_session)
        in: header
        name: Cookie
        type: string
      - description: State parameter
        in: query
        name: state
        type: string
      - description: Response type (code)
        in: query
        name: response_type
        type: string
      - description: Client ID
        in: query
        name: client_id
        type: string
      - description: Redirect URI
        in: query
        name: redirect_uri
        type: string
      - description: Scope
        in: query
        name: scope
        type: string
      - description: Nonce
        in: query
        name: nonce
        type: string
      - description: Code challenge
        in: query
        name: code_challenge
        type: string
      - description: Code challenge method
        in: query
        name: code_challenge_method
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect to client or login
          schema:
            type: string
      summary: Authorize endpoint (authorization code + PKCE)
      tags:
      - OIDC
...
  /introspect:
    post:
      consumes:
      - application/json
      - application/x-www-form-urlencoded
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.TokenIntrospectionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Token introspection endpoint (RFC 7662)
      tags:
      - OIDC
  /login:
    post:
      consumes:
      - application/json
      description: Sets the same iam_session cookie as OIDC so /authorize recognizes
        the browser without a second login. If the user has MFA (TOTP) enabled, `data`
        is MFALoginChallengeResponse (mfa_ticket) instead of LoginResponse; complete
        login with POST /mfa/challenge/verify. When the user belongs to multiple tenants
        and no tenant context is provided, `data` is TenantSelectionResponse.
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'When MFA is off: LoginResponse. When MFA is on: same envelope
            but `data` is MFALoginChallengeResponse (see top-level description).'
          schema:
            properties:
              data:
                $ref: '#/definitions/dtos.LoginResponse'
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              meta:
                $ref: '#/definitions/dtos.Meta'
            type: object
      summary: Login (dashboard/API) with email and password
      tags:
      - Auth
...
  /oidc/federation/{provider}/callback:
    get:
      parameters:
      - description: Provider id (e.g. google)
        in: path
        name: provider
        required: true
        type: string
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      - description: State
        in: query
        name: state
        required: true
        type: string
      responses:
        "302":
          description: Redirect to stored return_to
          schema:
            type: string
      summary: Federated OAuth callback
      tags:
      - OIDC
  /oidc/federation/{provider}/start:
    get:
      parameters:
      - description: Provider id (e.g. google)
        in: path
        name: provider
        required: true
        type: string
      - description: URL to return to after login (must be /authorize on this app)
        in: query
        name: return_to
        required: true
        type: string
      responses:
        "302":
          description: Redirect to identity provider
          schema:
            type: string
      summary: Start federated sign-in (OIDC browser flow)
      tags:
      - OIDC
  /oidc/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials + return_to
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.LoginRequest'
      produces:
      - application/json
      responses:
        "302":
          description: Redirect to return_to with session cookie set
          schema:
            type: string
      summary: Login (OIDC browser flow) and continue /authorize
      tags:
      - OIDC
...
  /userinfo:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Userinfo endpoint (Bearer access token from token endpoint, RS256)
      tags:
      - OIDC
...
schemes:
- http
- https
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Bearer Token Authentication. Use "Bearer {token}" as the value.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"